requirements.txt
Flask>=2.2
PyYAML>=6.0
playwright>=1.43
gunicorn